third_party/blink/web_tests/fast/dom/shadow/shadow-root-innerHTML-expected.txt

This tests the innerHTML property of a shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS 'innerHTML' in internals.shadowRoot is false
PASS root.innerHTML is defined.
PASS root.innerHTML is '<div id="div1"></div><div id="div2"></div>'
PASS root.innerHTML is 'Hello'
PASS root.innerHTML is '<p id="p3">HelloWorld</p>'
PASS root.firstChild.nodeName is 'P'
PASS root.querySelectorAll('div') is []
PASS root.getElementById('p3').innerHTML is 'HelloWorld'
PASS successfullyParsed is true

TEST COMPLETE
